requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
fastapi==0.85.1
filelock==3.8.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
h11==0.14.0
httplib2==0.20.4
httptools==0.5.0
idna==3.4
oauthlib==3.2.2
pbr==5.11.0
platformdirs==2.5.2
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
stevedore==4.1.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchfiles==0.18.0
websockets==10.4
